+++ e2e/test/hit-limit/test/loop.karma.spec.js
@@ -1,0 +1,7 @@
+describe('loop', () => {
+  it('should result in 15 for n=5 and a sum function', () => {
+    let result = 0;
+    loop(5, n => result += n);
+    expect(result).toBe(15);
+  });
+});