+++ packages/core/src/test-runner/test-runner-decorator.ts
@@ -1,46 +1,12 @@
 import { TestRunner, DryRunOptions, MutantRunOptions, MutantRunResult, DryRunResult } from '@stryker-mutator/api/test-runner';
-import { Disposable } from 'typed-inject';
 
-export class TestRunnerDecorator implements Required<TestRunner>, Disposable {
-  protected innerRunner!: TestRunner;
+import { ResourceDecorator } from '../concurrent';
 
-  constructor(private readonly testRunnerProducer: () => TestRunner) {
-    this.createInnerRunner();
-  }
-
-  public init(): Promise<void> {
-    if (this.innerRunner.init) {
-      return this.innerRunner.init();
-    } else {
-      return Promise.resolve();
-    }
-  }
-  protected createInnerRunner(): void {
-    this.innerRunner = this.testRunnerProducer();
-  }
-
+export class TestRunnerDecorator extends ResourceDecorator<TestRunner> {
   public dryRun(options: DryRunOptions): Promise<DryRunResult> {
-    return this.innerRunner.dryRun(options);
+    return this.innerResource.dryRun(options);
   }
   public mutantRun(options: MutantRunOptions): Promise<MutantRunResult> {
-    return this.innerRunner.mutantRun(options);
-  }
-
-  public dispose(): Promise<any> {
-    if (this.innerRunner.dispose) {
-      return this.innerRunner.dispose();
-    } else {
-      return Promise.resolve();
-    }
-  }
-
-  /**
-   * Disposes the current test runner and creates a new one
-   * To be used in decorators that need recreation.
-   */
-  protected async recover(): Promise<void> {
-    await this.dispose();
-    this.createInnerRunner();
-    return this.init();
+    return this.innerResource.mutantRun(options);
   }
 }