+++ packages/karma-runner/src/karma-plugins/stryker-mutant-coverage-adapter.ts
@@ -1,16 +1,23 @@
 // THIS FILE IS LOADED IN THE BROWSER AND SHOULD NOT BE A NODE / ES MODULE
 
-// @ts-expect-error
+// @ts-expect-error window is not defined
 const originalComplete = window.__karma__.complete.bind(window.__karma__);
-// @ts-expect-error
+// @ts-expect-error window is not defined
 window.__karma__.complete = (...args) => {
-  // @ts-expect-error
-  if (window.__strykerShouldReportCoverage__) {
+  function firstArg() {
     if (!args.length) {
       args.push({});
     }
-    // @ts-expect-error
-    args[0].mutantCoverage = window.__stryker__?.mutantCoverage;
+    return args[0];
+  }
+  // @ts-expect-error window is not defined
+  const context = window.__stryker__ || {};
+  // @ts-expect-error window is not defined
+  if (window.__strykerShouldReportCoverage__) {
+    firstArg().mutantCoverage = context.mutantCoverage;
+  }
+  if (context.hitCount !== undefined) {
+    firstArg().hitCount = context.hitCount;
   }
   originalComplete(...args);
 };