+++ packages/karma-runner/testResources/infinite-loop/infinite-loop.js
@@ -1,0 +1,8 @@
+function loop(n, action) {
+  let goOn = true;
+  while(goOn) {
+    action(n);
+    n--;
+    goOn = n > 0;
+  }
+}