+++ packages/mocha-runner/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.3.0](https://github.com/stryker-mutator/stryker-js/compare/v5.2.3...v5.3.0) (2021-08-07)
+
+
+### Features
+
+* **hit limit:** infinite loop prevention in karma-runner ([#3031](https://github.com/stryker-mutator/stryker-js/issues/3031)) ([fc732fc](https://github.com/stryker-mutator/stryker-js/commit/fc732fce8838c96be2fdf37aff69b12c996d7cb0))
+
+
+
+
+
+## [5.2.3](https://github.com/stryker-mutator/stryker-js/compare/v5.2.2...v5.2.3) (2021-08-01)
+
+
+### Bug Fixes
+
+* **mocha-runner:** clear error when require esm ([#3015](https://github.com/stryker-mutator/stryker-js/issues/3015)) ([a835f0b](https://github.com/stryker-mutator/stryker-js/commit/a835f0b57a9084b77a175b5eb14f409651c20c69)), closes [#3014](https://github.com/stryker-mutator/stryker-js/issues/3014)
+
+
+
+
+
 ## [5.2.2](https://github.com/stryker-mutator/stryker-js/compare/v5.2.1...v5.2.2) (2021-07-13)
 
 **Note:** Version bump only for package @stryker-mutator/mocha-runner